Refactor WebSocket Functionality into Organized Module Structure
- Moved WebSocket handling from the monolithic `websocket_wiki.py` file to a new `websocket/wiki_handler.py` module, enhancing modularity and maintainability.
- Preserved all existing functionality, including multi-provider AI support, RAG-powered document retrieval, and real-time streaming responses.
- Updated import statements across the codebase to reflect the new module structure, ensuring backward compatibility.
- Completed Phase 7.2 of the API restructure, establishing a solid foundation for the WebSocket layer and preparing for subsequent phases.
