+++ api/api/v1/chat.py
@@ -11,7 +11,7 @@ from typing import List, Optional
 
 # Import the actual chat implementations from the existing modules
 from ..simple_chat import chat_completions_stream, ChatCompletionRequest, ChatMessage
-from ..websocket_wiki import handle_websocket_chat
+from ..websocket import handle_websocket_chat
 
 # Create router for chat endpoints
 router = APIRouter()
