+++ memory-bank/activeContext.md
@@ -1,7 +1,7 @@
 # Active Context - DeepWiki Project
 
 ## Current Session Focus
-**API Restructure Phase 7.1** - 🟢 **COMPLETED** - Utilities successfully extracted and organized into comprehensive package with 83 functions across 6 modules.
+**API Restructure Phase 7.2** - 🟢 **COMPLETED** - WebSocket functionality successfully moved to organized module structure while preserving all existing functionality.
 
 ## Immediate Priorities
 1. ✅ **API Restructure Phase 1.1** - Directory structure and foundation completed
@@ -19,11 +19,12 @@
 13. ✅ **Phase 6.1 Data Processing** - **COMPLETED** - Data processors extracted and organized
 14. ✅ **Phase 6.2 Vector Operations** - **COMPLETED** - Vector operations and database management completed
 15. 🟢 **Phase 7.1 Utilities** - **COMPLETED** - Comprehensive utilities package created with 83 functions
+16. 🟢 **Phase 7.2 WebSocket** - **COMPLETED** - WebSocket functionality moved to organized module structure
 
 ## Current Work Context
 - **Phase**: API Restructure Implementation
-- **Focus Area**: Phase 7.1 - Utilities Extraction and Organization
-- **User Request**: "Execute task @TASK016-phase-7-1-utilities.md"
+- **Focus Area**: Phase 7.2 - WebSocket Functionality Relocation
+- **User Request**: "Execute task @TASK017-phase-7-2-websocket.md"
 - **Status**: 🟢 **COMPLETED** - 100% Complete
 
 ## Key Discoveries Made
@@ -84,36 +85,37 @@
 14. ✅ **Phase 5.3 App Configuration** - **COMPLETED** - FastAPI configuration extracted and organized
 15. ✅ **Phase 6.1 Data Processing** - **COMPLETED** - Data processors extracted and organized
 16. ✅ **Phase 6.2 Vector Operations** - **COMPLETED** - Vector operations and database management completed
-17. 🎯 **Ready for Phase 7.2** - Begin implementing WebSocket functionality
+17. 🟢 **Phase 7.2 WebSocket** - **COMPLETED** - WebSocket functionality moved to organized module structure
+18. 🎯 **Ready for Phase 7.3** - Begin implementing Prompts functionality
 
 ## Session Notes
-- **User**: Requested execution of TASK016 (Utilities)
-- **Approach**: Systematic extraction of utility functions from scattered codebase components into organized modules
-- **Focus**: Creating comprehensive utilities package with logical organization and comprehensive testing
-- **Status**: 🟢 **COMPLETED** - 100% Complete - All utilities extracted, organized, tested, and documented
+- **User**: Requested execution of TASK017 (WebSocket)
+- **Approach**: Systematic relocation of WebSocket functionality from monolithic file to organized module structure
+- **Focus**: Moving WebSocket functionality while preserving all existing features and updating imports
+- **Status**: 🟢 **COMPLETED** - 100% Complete - All WebSocket functionality moved, organized, and tested
 
 ## Context Preservation
-This session has successfully implemented Phase 7.1 of the API restructure, creating a comprehensive utilities package that consolidates 83 utility functions from scattered codebase components. The implemented utilities include:
+This session has successfully implemented Phase 7.2 of the API restructure, moving WebSocket functionality from the monolithic `websocket_wiki.py` file to the new organized `websocket/wiki_handler.py` module structure. The implemented WebSocket functionality includes:
 
-- **Text Processing Utilities**: Content analysis, detection (Markdown, YAML, JSON), readability metrics, and text processing
-- **File Operations Utilities**: Path handling, file system operations, repository utilities, and file validation
-- **Validation Utilities**: Input validation, format validation, and specialized validation functions
-- **Token Utilities**: Token counting, analysis, and text optimization for AI models
-- **Configuration Utilities**: Configuration loading, environment processing, and management functions
-- **Response Utilities**: AI provider response processing and normalization
+- **Multi-Provider AI Support**: Google, OpenAI, OpenRouter, Ollama, Azure, and Dashscope integration
+- **RAG-Powered Responses**: Document retrieval, context building, and intelligent responses
+- **Deep Research Functionality**: Multi-iteration research process with conversation continuity
+- **Real-Time Communication**: WebSocket streaming responses and connection management
+- **File Integration**: File content retrieval and context-aware responses
+- **Error Handling**: Comprehensive error handling and fallback mechanisms
 
-All utilities have been extracted, organized into logical modules, comprehensively tested, and documented. The utilities package is now production-ready and provides significant improvement in discoverability, reusability, and maintainability over the scattered implementation. The system is ready for Phase 7.2, which will focus on WebSocket functionality.
+All WebSocket functionality has been preserved, organized into a clean module structure, and imports have been updated throughout the codebase. The WebSocket module now provides a clean, organized interface for real-time communication while maintaining full backward compatibility. The system is ready for Phase 7.3, which will focus on Prompts functionality.
 
 ## Technical Achievements
-1. **Utilities Package Architecture**: Successfully created comprehensive utilities package with 6 logical modules
-2. **Function Extraction**: Extracted 83 utility functions from scattered codebase components
-3. **Modular Organization**: Organized utilities into logical modules for easy discovery and maintenance
-4. **Comprehensive Testing**: Implemented test suite with 100% pass rate for all utility functions
-5. **Documentation**: Created detailed README with usage examples and best practices
-6. **Type Safety**: Added comprehensive type hints for all utility functions
-7. **Error Handling**: Implemented robust error handling throughout all utility functions
-8. **Backward Compatibility**: Maintained all existing functionality while improving organization
-9. **Import Resolution**: Resolved complex import path issues during extraction
-10. **Quality Assurance**: Established utilities package as production-ready foundation
+1. **WebSocket Module Architecture**: Successfully created organized WebSocket module structure
+2. **Functionality Preservation**: Maintained all existing WebSocket functionality including multi-provider AI support
+3. **Module Organization**: Organized WebSocket functionality into clean, maintainable module structure
+4. **Import Management**: Updated all import statements throughout the codebase to use new module structure
+5. **Backward Compatibility**: Ensured all existing functionality continues to work without changes
+6. **Code Organization**: Eliminated monolithic file structure in favor of organized modules
+7. **Module Exports**: Properly configured module exports for clean import interface
+8. **File Cleanup**: Removed old websocket_wiki.py file after successful migration
+9. **Testing Validation**: Verified successful imports and module functionality
+10. **Quality Assurance**: Established WebSocket module as production-ready foundation
 
-The utilities package is now fully implemented and production-ready, providing significant improvement in discoverability, reusability, and maintainability over the scattered implementation. This establishes a solid foundation for the utilities layer and prepares the system for the next phase of the API restructure.
+The WebSocket functionality is now fully organized into a clean module structure while maintaining all existing features. This establishes a solid foundation for the WebSocket layer and prepares the system for the next phase of the API restructure.
