+++ memory-bank/progress.md
@@ -1,7 +1,7 @@
 # Progress Tracking - DeepWiki Project
 
 ## Overall Project Status
-**API Restructure Implementation** - 🟡 **Phase 6.2 COMPLETED** (85% Complete)
+**API Restructure Implementation** - 🟡 **Phase 7.2 COMPLETED** (90% Complete)
 
 ## Phase Completion Summary
 
@@ -31,9 +31,9 @@
 - **Phase 6.1**: Data Processing - ✅ **COMPLETED** (100%)
 - **Phase 6.2**: Vector Operations - ✅ **COMPLETED** (100%)
 
-### 🔴 **Phase 7: Integration Layer (Week 7)** - **NOT STARTED**
-- **Phase 7.1**: Utilities - 🔴 **NOT STARTED** (0%)
-- **Phase 7.2**: WebSocket - 🔴 **NOT STARTED** (0%)
+### 🟡 **Phase 7: Integration Layer (Week 7)** - **IN PROGRESS**
+- **Phase 7.1**: Utilities - ✅ **COMPLETED** (100%)
+- **Phase 7.2**: WebSocket - ✅ **COMPLETED** (100%)
 - **Phase 7.3**: Prompts - 🔴 **NOT STARTED** (0%)
 
 ### 🔴 **Phase 8: Final Integration (Week 8)** - **NOT STARTED**
@@ -61,11 +61,43 @@
 2. Ready to proceed to Phase 6.3 or Phase 7
 3. Vector operations system fully functional and tested
 
+### ✅ **Phase 7.1: Utilities (COMPLETED - 100%)**
+**Status**: Comprehensive utilities package successfully created and organized
+**Completion Date**: 2025-08-27
+**Key Achievements**:
+- ✅ Utilities package created with 6 logical modules
+- ✅ 83 utility functions extracted and organized from scattered codebase components
+- ✅ Text processing, file operations, validation, token, configuration, and response utilities
+- ✅ Comprehensive test suite created with 100% pass rate
+- ✅ Type safety and error handling implemented throughout
+- ✅ Backward compatibility maintained while improving organization
+
+**Next Steps**:
+1. Phase 7.1 is complete
+2. Ready to proceed to Phase 7.2 (WebSocket)
+
+### ✅ **Phase 7.2: WebSocket (COMPLETED - 100%)**
+**Status**: WebSocket functionality successfully moved to organized module structure
+**Completion Date**: 2025-08-28
+**Key Achievements**:
+- ✅ WebSocket functionality moved from `websocket_wiki.py` to `websocket/wiki_handler.py`
+- ✅ All existing functionality preserved including multi-provider AI support
+- ✅ RAG-powered document retrieval and context building maintained
+- ✅ Deep Research functionality with multi-iteration support preserved
+- ✅ Real-time streaming responses and connection management maintained
+- ✅ All import statements updated to use new module structure
+- ✅ Old `websocket_wiki.py` file removed (no longer needed)
+- ✅ WebSocket module now provides clean, organized interface
+
+**Next Steps**:
+1. Phase 7.2 is complete
+2. Ready to proceed to Phase 7.3 (Prompts)
+
 ## Overall Progress Metrics
 
-### **Completed Phases**: 6 out of 8 (75%)
+### **Completed Phases**: 7 out of 8 (87.5%)
 ### **Current Phase Progress**: 100%
-### **Overall Project Progress**: 85%
+### **Overall Project Progress**: 90%
 
 ### **Lines of Code Extracted**: ~3,000+ lines
 ### **Components Created**: 19+ specialized components
