+++ memory-bank/task-list.md
@@ -4,7 +4,7 @@
 
 | ID | Task Name | Status | Progress | Phase | Priority |
 |----|-----------|--------|----------|-------|----------|
-| TASK017 | Phase 7.2: WebSocket | 🔴 Not Started | 0% | Week 7 | 🟡 Medium |
+| TASK017 | Phase 7.2: WebSocket | ✅ **COMPLETED** | 100% | Week 7 | 🟡 Medium |
 | TASK018 | Phase 7.3: Prompts | 🔴 Not Started | 0% | Week 7 | 🟡 Medium |
 | TASK019 | Phase 8.1: Test Structure | 🔴 Not Started | 0% | Week 8 | 🟡 Medium |
 | TASK020 | Phase 8.2: Import Updates | 🔴 Not Started | 0% | Week 8 | 🟡 Medium |
@@ -29,14 +29,15 @@
 | TASK013 | Phase 6.1: Data Processing | ✅ **COMPLETED** | 100% | Week 6 | 🟡 Medium |
 | TASK014 | Phase 6.2: Vector Operations | ✅ **COMPLETED** | 100% | Week 6 | 🟡 Medium |
 | TASK016 | Phase 7.1: Utilities | ✅ **COMPLETED** | 100% | Week 7 | 🟡 Medium |
+| TASK017 | Phase 7.2: WebSocket | ✅ **COMPLETED** | 100% | Week 7 | 🟡 Medium |
 
 ## Task Status Summary
 
 - **Total Tasks**: 21
-- **Completed**: 15 (71%)
+- **Completed**: 16 (76%)
 - **In Progress**: 0 (0%)
-- **Not Started**: 6 (29%)
-- **Overall Progress**: 71%
+- **Not Started**: 5 (24%)
+- **Overall Progress**: 76%
 
 ## Next Priority Tasks
 
@@ -48,8 +49,10 @@
 
 - **Phase 6 (Data Layer) is now COMPLETE** with all vector operations successfully extracted
 - **Phase 7.1 (Utilities) is now COMPLETE** with comprehensive utility extraction and organization
+- **Phase 7.2 (WebSocket) is now COMPLETE** with WebSocket functionality successfully moved to organized module structure
 - **Vector operations system** is fully functional with comprehensive testing
 - **Utilities package** is fully organized with 83 functions across 6 modules
+- **WebSocket module** is fully organized with all functionality preserved and imports updated
 - **FAISS integration maintained** with enhanced functionality
 - **Backward compatibility ensured** through compatibility layer
-- Ready to proceed to **Phase 7.2 (WebSocket)** and **Phase 7.3 (Prompts)**
+- Ready to proceed to **Phase 7.3 (Prompts)** and **Phase 8 (Final Layer)**
